- Fix Win64 build. - Style cleanup: use Page::is_valid() instead of NULL-check, use Heap::CreateFillerObjectAt() instead of dummy free list nodes Review URL: http://codereview.chromium.org/1691009
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@4484 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.h
@@ -641,7 +641,7 @@ class MemoryAllocator : public AllStatic {
 
   static Page* RelinkPagesInChunk(int chunk_id,
                                   Address chunk_start,
-                                  int chunk_size,
+                                  size_t chunk_size,
                                   Page* prev,
                                   Page** last_page_in_use);
 };
